chore: Remove util.rs; use Program directly
