+++ rex-engine/src/lib.rs
@@ -4,4 +4,3 @@ pub mod error;
 pub mod eval;
 pub mod ftable;
 pub mod program;
-pub mod util;
